requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==23.0.0
playwright==1.58.0
playwright-stealth==2.0.3
python-dotenv==1.0.0
filelock==3.25.2
tenacity==9.1.2
